Added Log-in screen and Site Administration
Log-in screen is prompted everytime there is not a session open.
Site Administration is used to create new users and display the existing
users on the database. User can also update his/her password and display name.

A Form template and Input template .erb files and classes have been
created in order to reuse and abstract the form components. Error messages
are properly displayed and handled.