Replaced 'change password' with JavaScript
Removed old change password screen and replaced it with Backbone templates
and JS. Removed all the old forms and form inputs classes (we can bring it
back later... but for now they are gone).

Fixed issue with redirection not working and an issue with error message
showing up even when trying the base path.

Success messages have been added to user management