Finalized common API for :things
Added the get/post/put/delete actions for 'things'. 'Things' is the
abstraction of any model that corresponds to a table on the database. The
initial endpoint should provide a URL, an array of column names and an
array of keys (+ some other titles for the forms) and everything else goes
to the same codepath (creating a new item/thing, updating, deleting).

On the next iterations of the code we should improve this so that the
model itself is the one providing the implementation instead of writing
it directly in a route.

A second improvement must be done so that we can show the relationships
between tables automatically. e.g. a many_to_one relationship on a
database table might be able to be represented like a dropdown or
something like that.

This is a first version that 'works', but not what we want at the end